Move error handling out of state creation mapping
This makes sure the component is not submitted in case of an error.

COSDK-1120
