+++ card/src/main/java/com/adyen/checkout/card/internal/ui/StoredCardComponent.kt
@@ -14,35 +14,46 @@ import com.adyen.checkout.card.internal.data.model.Brand
 import com.adyen.checkout.card.internal.data.model.DetectedCardType
 import com.adyen.checkout.card.internal.ui.model.CardComponentParams
 import com.adyen.checkout.card.internal.ui.model.StoredCVCVisibility
+import com.adyen.checkout.card.internal.ui.state.CardPaymentComponentState
+import com.adyen.checkout.card.internal.ui.state.StoredCardComponentState
 import com.adyen.checkout.card.internal.ui.state.StoredCardComponentStateFactory
 import com.adyen.checkout.card.internal.ui.state.StoredCardComponentStateReducer
 import com.adyen.checkout.card.internal.ui.state.StoredCardComponentStateValidator
 import com.adyen.checkout.card.internal.ui.state.StoredCardIntent
 import com.adyen.checkout.card.internal.ui.state.StoredCardViewStateProducer
-import com.adyen.checkout.card.internal.ui.state.toPaymentComponentState
 import com.adyen.checkout.card.internal.ui.view.StoredCardContent
 import com.adyen.checkout.core.analytics.internal.AnalyticsManager
 import com.adyen.checkout.core.analytics.internal.ErrorEvent
 import com.adyen.checkout.core.analytics.internal.GenericEvents
 import com.adyen.checkout.core.common.CardBrand
 import com.adyen.checkout.core.common.CardType
+import com.adyen.checkout.core.common.helper.runCompileOnly
 import com.adyen.checkout.core.common.internal.helper.bufferedChannel
+import com.adyen.checkout.core.components.data.PaymentComponentData
 import com.adyen.checkout.core.components.data.model.paymentmethod.StoredCardPaymentMethod
 import com.adyen.checkout.core.components.internal.PaymentComponentEvent
 import com.adyen.checkout.core.components.internal.data.provider.SdkDataProvider
 import com.adyen.checkout.core.components.internal.ui.PaymentComponent
 import com.adyen.checkout.core.components.internal.ui.state.ComponentStateFlow
 import com.adyen.checkout.core.components.internal.ui.state.model.RequirementPolicy
+import com.adyen.checkout.core.components.internal.ui.state.model.getPaymentDataValue
 import com.adyen.checkout.core.components.internal.ui.state.viewState
+import com.adyen.checkout.core.components.paymentmethod.CardDetails
+import com.adyen.checkout.core.error.internal.GenericError
 import com.adyen.checkout.core.error.internal.InternalCheckoutError
+import com.adyen.checkout.cse.EncryptedCard
 import com.adyen.checkout.cse.EncryptionException
+import com.adyen.checkout.cse.UnencryptedCard
 import com.adyen.checkout.cse.internal.BaseCardEncryptor
+import com.adyen.threeds2.ThreeDS2Service
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.receiveAsFlow
 
+@Suppress("TooManyFunctions")
+internal class StoredCardComponent
 @Suppress("LongParameterList")
-internal class StoredCardComponent(
+constructor(
     private val storedPaymentMethod: StoredCardPaymentMethod,
     private val analyticsManager: AnalyticsManager,
     private val cardEncryptor: BaseCardEncryptor,
@@ -104,20 +115,62 @@ internal class StoredCardComponent(
 
     override fun submit() {
         if (componentStateValidator.isValid(componentState.value)) {
-            val paymentComponentState = componentState.value.toPaymentComponentState(
-                cardEncryptor = cardEncryptor,
+            val encryptedCard = componentState.value.encryptCard() ?: return
+            val paymentComponentState = createPaymentComponentState(
+                encryptedCard = encryptedCard,
                 sdkDataProvider = sdkDataProvider,
                 storedPaymentMethodId = storedPaymentMethod.id,
                 paymentMethodType = storedPaymentMethod.type,
-                onEncryptionFailed = ::onEncryptionError,
-                onPublicKeyNotFound = ::onPublicKeyNotFound,
-                publicKey = publicKey,
             )
             val event = PaymentComponentEvent.Submit(paymentComponentState)
             eventChannel.trySend(event)
         } else {
             onIntent(StoredCardIntent.HighlightValidationErrors)
         }
+    }
+
+    private fun StoredCardComponentState.encryptCard(): EncryptedCard? {
+        publicKey ?: run {
+            onPublicKeyNotFound()
+            return null
+        }
+        return try {
+            val unencryptedCardBuilder = UnencryptedCard.Builder()
+            securityCode.getPaymentDataValue()?.let {
+                unencryptedCardBuilder.setCvc(it)
+            }
+            cardEncryptor.encryptFields(unencryptedCardBuilder.build(), publicKey)
+        } catch (e: EncryptionException) {
+            onEncryptionError(e)
+            null
+        }
+    }
+
+    private fun createPaymentComponentState(
+        encryptedCard: EncryptedCard,
+        sdkDataProvider: SdkDataProvider,
+        storedPaymentMethodId: String,
+        paymentMethodType: String,
+    ): CardPaymentComponentState {
+        val cardDetails = CardDetails(
+            type = paymentMethodType,
+            sdkData = sdkDataProvider.createEncodedSdkData(
+                threeDS2SdkVersion = runCompileOnly { ThreeDS2Service.INSTANCE.sdkVersion }.getOrNull(),
+            ),
+            storedPaymentMethodId = storedPaymentMethodId,
+            encryptedSecurityCode = encryptedCard.encryptedSecurityCode,
+        )
+
+        val paymentComponentData = PaymentComponentData(
+            paymentMethod = cardDetails,
+            storePaymentMethod = null,
+            order = null,
+        )
+
+        return CardPaymentComponentState(
+            data = paymentComponentData,
+            isValid = true,
+        )
     }
 
     override fun requiresUserInteraction(): Boolean =
@@ -127,25 +180,29 @@ internal class StoredCardComponent(
         onIntent(StoredCardIntent.UpdateLoading(isLoading))
     }
 
-    override fun onCleared() = Unit
-
     private fun onIntent(intent: StoredCardIntent) {
         componentState.handleIntent(intent)
     }
 
-    @Suppress("UNUSED_PARAMETER")
     private fun onEncryptionError(e: EncryptionException) {
         val event = GenericEvents.error(storedPaymentMethod.type, ErrorEvent.ENCRYPTION)
         analyticsManager.trackEvent(event)
-        // exceptionChannel.trySend(e)
+        emitError(GenericError("Encryption error", e))
     }
 
-    @Suppress("UNUSED_PARAMETER")
-    private fun onPublicKeyNotFound(e: InternalCheckoutError) {
+    private fun onPublicKeyNotFound() {
         val event = GenericEvents.error(storedPaymentMethod.type, ErrorEvent.API_PUBLIC_KEY)
         analyticsManager.trackEvent(event)
-        // exceptionChannel.trySend(e)
+        emitError(GenericError("Public key is missing."))
     }
+
+    private fun emitError(error: InternalCheckoutError) {
+        eventChannel.trySend(
+            PaymentComponentEvent.Error(error),
+        )
+    }
+
+    override fun onCleared() = Unit
 
     companion object {
         private val NO_CVC_BRANDS: Set<CardBrand> = setOf(CardBrand(txVariant = CardType.BCMC.txVariant))
